api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
anthropic==0.75.0
python-dotenv==1.0.1
